add getStyle
