+++ packages/sdks/src/blocks/button/button.lite.tsx
@@ -38,6 +38,7 @@ export default function Button(props: ButtonProps) {
         svelte: filterAttrs(props.attributes, 'on:', true),
         default: {},
       })}
+      style={{}}
     >
       {props.text}
     </DynamicRenderer>
