+++ packages/sdks/src/components/block/components/block-wrapper.lite.tsx
@@ -3,6 +3,7 @@ import { useMetadata } from '@builder.io/mitosis';
 import type { BuilderContextInterface } from '../../../context/types.js';
 import { getBlockActions } from '../../../functions/get-block-actions.js';
 import { getBlockProperties } from '../../../functions/get-block-properties.js';
+import { getStyle } from '../../../functions/get-style.js';
 import type { BuilderBlock } from '../../../types/builder-block.js';
 import type { PropsWithChildren } from '../../../types/typescript.js';
 import DynamicRenderer from '../../dynamic-renderer/dynamic-renderer.lite.jsx';
@@ -35,6 +36,11 @@ export default function BlockWrapper(
     <DynamicRenderer
       TagName={props.Wrapper}
       attributes={getBlockProperties({
+        block: props.block,
+        context: props.context.value,
+        includeStyles: false,
+      })}
+      style={getStyle({
         block: props.block,
         context: props.context.value,
       })}
