+++ packages/sdks/src/components/block/components/component-ref/component-ref.helpers.ts
@@ -49,6 +49,7 @@ export const getWrapperProps = ({
           attributes: getBlockProperties({
             block: builderBlock,
             context: contextValue,
+            includeStyles: true,
           }),
         }
       : {}),
